Ref #12 Desktopアプリも分離
